+++ src/runtime/component/instance.ts
@@ -38,6 +38,7 @@ import {
   getOwnershipSignal,
   OwnershipRecord,
   attachOwnership,
+  invalidatePendingChecks,
 } from '../ownership/record';
 import { runScheduledComponent } from './commit';
 import {
@@ -125,6 +126,7 @@ function enqueueComponentRun(this: ComponentInstance): void {
   }
 
   this.hasPendingUpdate = true;
+  invalidatePendingChecks();
   enqueueRuntimeTask(ensurePendingRunTask(this));
 }
 
@@ -415,6 +417,7 @@ export function executeComponent(instance: ComponentInstance): void {
 
   // Initial renders use the same cancellable task as state-driven rerenders.
   instance.hasPendingUpdate = true;
+  invalidatePendingChecks();
   enqueueRuntimeTask(ensurePendingRunTask(instance));
 }